우창하 : 배경이미지가 프리카메라 가리던 현상 수정

+++ DirectX_MapleStory/GameEngineContents/RenderActor.cpp
@@ -11,12 +11,12 @@ RenderActor::~RenderActor()
 
 }
 
-void RenderActor::Init(int _Order, int _Depth)
+void RenderActor::Init(int _Order, float _Depth)
 {
 	if (nullptr == Renderer)
 	{
 		Renderer = CreateComponent<GameEngineSpriteRenderer>(_Order);
-		Renderer->Transform.SetLocalPosition({ 0, 0, static_cast<float>(_Depth) });
+		Renderer->Transform.SetLocalPosition({ 0, 0, _Depth });
 		Renderer->AutoSpriteSizeOn();
 	}
 }
